+++ frontend/src/Editor/RealtimeEditor.jsx
@@ -11,7 +11,7 @@ const ydoc = new Y.Doc();
 const getWebsocketUrl = () => {
   const protocol = window.location.protocol === 'https:' ? 'wss' : 'ws';
   const re = /https?:\/\//g;
-  if (re.test(config.apiUrl)) return `${protocol}://${config.websocketUrl.replace(/(^\w+:|^)\/\//, '')}`;
+  if (re.test(config.apiUrl)) return `${protocol}://${config.apiUrl.replace(/(^\w+:|^)\/\//, '').replace('/api', '')}`;
 
   return `${protocol}://${window.location.host}`;
 };
